API change for more clean usage of selector descriptors
Annotator's `onSuccessfulSelection()` function now expects to
find `segments` inside the passed event, instead of `targets`.

The creation of targets is no longer the responsibility of the
UI code which detected the selection; targets will automatically
be created my calling the registered selector descriptors on
the passed segments.

Updated the TextAnchors plugin for the new UI.